server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["nnniv <nivroh2016@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sanic = "^24.6.0"
mysql-connector-python = "^9.0.0"
mariadb = "1.1.9"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
ngrok = "^1.4.0"
sanic-ext = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
